scraping/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
pdfplumber>=0.11
playwright>=1.40
supabase>=2.0
pydantic>=2.5
python-dotenv>=1.0
numpy>=1.26
pandas>=2.1
rapidfuzz>=3.5
